backend/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-multipart==0.0.20
apscheduler==3.10.4
pyyaml>=6.0.1
pandas>=2.2.0
openpyxl>=3.1.2
jinja2>=3.1.3
python-dotenv>=1.0.1
